airi/services/vision-fastvm/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12

# Vision and ML
torch>=2.0.0
transformers>=4.40.0
Pillow==11.0.0
accelerate==0.27.0

# Screen capture
mss==9.0.2
pyautogui==0.9.54

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
numpy==1.26.4

# Monitoring
prometheus-client==0.21.0